+++ src/deps.js
@@ -1,3 +1,6 @@
 
 export { default as fs } from 'fs';
 export { default as path } from 'path';
+export { default as stream } from 'stream';
+export { default as util } from 'util';
+export { default as TS } from 'textics-stream';